add restrictions for Prow staging process
